Revert "Ft Borrow book request and accept  request 152996871"
